Fix a typo in the package name in the integration-test module
