+++ tests/src/test/java/org/apache/camel/kafkaconnector/DefaultKafkaConnectPropertyFactory.java
@@ -16,15 +16,18 @@
  *
  */
 
-package org.apache.camel.kakfaconnector.clients.kafka;
+package org.apache.camel.kafkaconnector;
 
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
+import org.apache.kafka.connect.runtime.standalone.StandaloneConfig;
 
 import java.util.Properties;
-import java.util.UUID;
 
-public class DefaultProducerPropertyFactory implements ProducerPropertyFactory {
+
+/**
+ * A set of properties for the Kafka connect runtime that match the standard configuration
+ * used for the standalone CLI connect runtime.
+ */
+public class DefaultKafkaConnectPropertyFactory implements KafkaConnectPropertyFactory {
     private final String bootstrapServer;
 
     /**
@@ -32,7 +35,7 @@ public class DefaultProducerPropertyFactory implements ProducerPropertyFactory {
      * @param bootstrapServer the address of the server in the format
      *                       PLAINTEXT://${address}:${port}
      */
-    public DefaultProducerPropertyFactory(String bootstrapServer) {
+    public DefaultKafkaConnectPropertyFactory(String bootstrapServer) {
         this.bootstrapServer = bootstrapServer;
     }
 
@@ -40,12 +43,13 @@ public class DefaultProducerPropertyFactory implements ProducerPropertyFactory {
     public Properties getProperties() {
         Properties props = new Properties();
 
-        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServer);
-        props.put(ProducerConfig.CLIENT_ID_CONFIG, UUID.randomUUID().toString());
-        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG,
-                StringSerializer.class.getName());
-        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG,
-                StringSerializer.class.getName());
+        props.put(StandaloneConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServer);
+        props.put(StandaloneConfig.KEY_CONVERTER_CLASS_CONFIG, "org.apache.kafka.connect.json.JsonConverter");
+        props.put(StandaloneConfig.VALUE_CONVERTER_CLASS_CONFIG, "org.apache.kafka.connect.json.JsonConverter");
+        props.put(StandaloneConfig.OFFSET_STORAGE_FILE_FILENAME_CONFIG, this.getClass().getResource("/").getPath() + "connect.offsets");
+        props.put(StandaloneConfig.OFFSET_COMMIT_INTERVAL_MS_CONFIG, "10000");
+        props.put(StandaloneConfig.PLUGIN_PATH_CONFIG, "");
+        props.put(StandaloneConfig.REST_PORT_CONFIG, "9999");
 
         return props;
     }
