+++ tests/src/test/java/org/apache/camel/kafkaconnector/KafkaConnectPropertyFactory.java
@@ -16,37 +16,21 @@
  *
  */
 
-package org.apache.camel.kakfaconnector;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+package org.apache.camel.kafkaconnector;
 
 import java.util.Properties;
 
-import static junit.framework.TestCase.fail;
 
 /**
- * Common test constants and utilities
+ * An interface for producing different types of Kafka connect properties. The CLI runtime
+ * equivalent for this file is the connect-standalone.properties
  */
-public final class TestCommon {
-    private static final Logger log = LoggerFactory.getLogger(TestCommon.class);
-
-    private TestCommon() {}
+public interface KafkaConnectPropertyFactory {
 
     /**
-     * The default topic for usage during the tests
+     * Gets the properties used to configure the Kafka connect runtime
+     * @return a Properties object containing the set of properties for the Kafka connect
+     * runtime
      */
-    public static final String DEFAULT_TEST_TOPIC = "mytopic";
-
-    /**
-     * The default JMS queue name used during the tests
-     */
-    public static final String DEFAULT_JMS_QUEUE = "ckc.queue";
-
-
-    public static void failOnConnectorError(Throwable error, Properties connectorProps, String name) {
-        log.error("Failed to create job for {} with properties", name, connectorProps,
-                error);
-        fail("Failed to create job for " + name);
-    }
+    Properties getProperties();
 }
