+++ tests/src/test/java/org/apache/camel/kafkaconnector/sink/jms/CamelJMSPropertyFactory.java
@@ -16,35 +16,45 @@
  *
  */
 
-package org.apache.camel.kakfaconnector.source.timer;
+package org.apache.camel.kafkaconnector.sink.jms;
 
-import org.apache.camel.kakfaconnector.ConnectorPropertyFactory;
+import org.apache.camel.kafkaconnector.ConnectorPropertyFactory;
 import org.apache.kafka.connect.runtime.ConnectorConfig;
 
 import java.util.Properties;
 
-class CamelTimerPropertyFactory implements ConnectorPropertyFactory {
+
+/**
+ * Creates the set of properties used by a Camel JMS Sink Connector
+ */
+class CamelJMSPropertyFactory implements ConnectorPropertyFactory {
     private final int tasksMax;
     private final String topic;
-    private int repeatCount;
+    private final String queue;
 
-    public CamelTimerPropertyFactory(int tasksMax, String topic, int repeatCount) {
+    private final String brokerURL;
+
+    CamelJMSPropertyFactory(int tasksMax, String topic, String queue, String brokerURL) {
         this.tasksMax = tasksMax;
         this.topic = topic;
-        this.repeatCount = repeatCount;
+        this.queue = queue;
+        this.brokerURL = brokerURL;
     }
 
     @Override
     public Properties getProperties() {
         Properties connectorProps = new Properties();
-        connectorProps.put(ConnectorConfig.NAME_CONFIG, "CamelSourceConnector");
+        connectorProps.put(ConnectorConfig.NAME_CONFIG, "CamelJmsSinkConnector");
         connectorProps.put("tasks.max", String.valueOf(tasksMax));
 
-        connectorProps.put(ConnectorConfig.CONNECTOR_CLASS_CONFIG, "org.apache.camel.kafkaconnector.CamelSourceConnector");
+        connectorProps.put(ConnectorConfig.CONNECTOR_CLASS_CONFIG, "org.apache.camel.kafkaconnector.CamelSinkConnector");
         connectorProps.put(ConnectorConfig.KEY_CONVERTER_CLASS_CONFIG, "org.apache.kafka.connect.storage.StringConverter");
         connectorProps.put(ConnectorConfig.VALUE_CONVERTER_CLASS_CONFIG, "org.apache.kafka.connect.storage.StringConverter");
-        connectorProps.put("camel.source.url", String.format("timer:kafkaconnector?repeatCount=%d", repeatCount));
-        connectorProps.put("camel.source.kafka.topic", topic);
+        connectorProps.put("camel.sink.url", "sjms2://queue:" + queue);
+        connectorProps.put("topics", topic);
+
+        connectorProps.put("camel.component.sjms2.connection-factory", "#class:org.apache.activemq.ActiveMQConnectionFactory");
+        connectorProps.put("camel.component.sjms2.connection-factory.brokerURL", brokerURL);
 
         return connectorProps;
     }
