+++ tests/src/test/java/org/apache/camel/kafkaconnector/source/timer/CamelTimerPropertyFactory.java
@@ -16,44 +16,35 @@
  *
  */
 
-package org.apache.camel.kakfaconnector.source.jms;
+package org.apache.camel.kafkaconnector.source.timer;
 
-import org.apache.camel.kakfaconnector.ConnectorPropertyFactory;
+import org.apache.camel.kafkaconnector.ConnectorPropertyFactory;
 import org.apache.kafka.connect.runtime.ConnectorConfig;
 
 import java.util.Properties;
 
-
-/**
- * Creates the set of properties used by a Camel JMS Sink Connector
- */
-class CamelJMSPropertyFactory implements ConnectorPropertyFactory {
+class CamelTimerPropertyFactory implements ConnectorPropertyFactory {
     private final int tasksMax;
     private final String topic;
-    private final String queue;
+    private int repeatCount;
 
-    private final String brokerURL;
-
-    CamelJMSPropertyFactory(int tasksMax, String topic, String queue, String brokerURL) {
+    public CamelTimerPropertyFactory(int tasksMax, String topic, int repeatCount) {
         this.tasksMax = tasksMax;
         this.topic = topic;
-        this.queue = queue;
-        this.brokerURL = brokerURL;
+        this.repeatCount = repeatCount;
     }
 
     @Override
     public Properties getProperties() {
         Properties connectorProps = new Properties();
-        connectorProps.put(ConnectorConfig.NAME_CONFIG, "CamelJMSSourceConnector");
+        connectorProps.put(ConnectorConfig.NAME_CONFIG, "CamelSourceConnector");
         connectorProps.put("tasks.max", String.valueOf(tasksMax));
 
         connectorProps.put(ConnectorConfig.CONNECTOR_CLASS_CONFIG, "org.apache.camel.kafkaconnector.CamelSourceConnector");
         connectorProps.put(ConnectorConfig.KEY_CONVERTER_CLASS_CONFIG, "org.apache.kafka.connect.storage.StringConverter");
         connectorProps.put(ConnectorConfig.VALUE_CONVERTER_CLASS_CONFIG, "org.apache.kafka.connect.storage.StringConverter");
-        connectorProps.put("camel.source.url", "sjms2://queue:" + queue);
+        connectorProps.put("camel.source.url", String.format("timer:kafkaconnector?repeatCount=%d", repeatCount));
         connectorProps.put("camel.source.kafka.topic", topic);
-        connectorProps.put("camel.component.sjms2.connection-factory", "#class:org.apache.activemq.ActiveMQConnectionFactory");
-        connectorProps.put("camel.component.sjms2.connection-factory.brokerURL", brokerURL);
 
         return connectorProps;
     }
