+++ tests/src/test/resources/org/apache/camel/kafkaconnector/Dockerfile
